chore/ Moved server logic into top level
